face-recognition-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
dlib==19.22.99